* FIX: Deleted 'const' on extra_day_count * FIX/DISABLE: Fixed or disabled warnings on xrCDB * DEL: Removed redundant includes

+++ code/engine.vc2008/xrCDB/xr_collide_defs.h
@@ -128,7 +128,7 @@ namespace collide
 			if (0==results.capacity())	results.reserve(8);
 			results.push_back			(res);
 		}
-		IC int			r_count			()	{ return results.size();	}
+		IC size_t		r_count			()	{ return results.size();	}
 		IC rq_result*	r_begin			()	{ return &*results.begin();	}
 		IC rq_result*	r_end			()	{ return &*results.end();	}
 		IC void			r_clear			()	{ results.clear();	}
